Import of MATIU/SQL-Bibliosoph-2.40 from CPAN.
gitpan-cpan-distribution: SQL-Bibliosoph
gitpan-cpan-version:      2.40
gitpan-cpan-path:         MATIU/SQL-Bibliosoph-2.40.tar.gz
gitpan-cpan-author:       MATIU
gitpan-cpan-maturity:     released

+++ lib/SQL/Bibliosoph.pm
@@ -10,7 +10,7 @@ package SQL::Bibliosoph; {
     use SQL::Bibliosoph::Query;
     use SQL::Bibliosoph::CatalogFile;
 
-    our $VERSION = "2.30";
+    our $VERSION = "2.40";
 
 
     has 'dbh'       => ( is => 'ro', isa => 'DBI::db',  required=> 1);
@@ -116,15 +116,18 @@ package SQL::Bibliosoph; {
         if ( $self->memc() ) {
             $self->d("Expiring group $group\n");
 
-
-            my $md5s = $self->memc()->get($group . '-g');
-            foreach (split /:/, $md5s) {
-                next if ! $_;
             
+            if (my $md5s = $self->memc()->get($group . '-g') ) {
+
+
+                foreach (split /:/, $md5s) {
+                    next if ! $_;
 #$self->d("\t\t expiring query in group $group : $_");
 
-                $self->memc()->delete($_);
+                    $self->memc()->delete($_);
+                }
             }
+
             $self->memc()->delete( $group . '-g' );
         }
         else {
@@ -136,13 +139,24 @@ package SQL::Bibliosoph; {
     sub add_to_group {
         my ($self, $group, $md5, $md5c) = @_;
 
+
         $group = $group . '-g';
+
+        my $all_md5s = $self->memc()->get( $group);
+
+        if ( $all_md5s && index($all_md5s, $md5) >= 0 ) {
+#$self->d("\t\t already stored:g:  ".$group ." md5: $md5");
+                return;
+        }
+
+
         $md5 .= ':' . $md5c if $md5c;
         $md5 .= ':';
 
         $self->memc()->append( $group, $md5) 
             || $self->memc()->set( $group, $md5)
             ;
+
 
 #$self->d("\t\t storing query in group ".$group ."$md5");
     }
@@ -249,7 +263,7 @@ package SQL::Bibliosoph; {
 
 
                     ## check memcached
-                    my $md5 = md5_hex( join ('', $name, map { $_ // 'NULL'  } @_ ));
+                    my $md5 = md5_hex( join (',', $name, map { $_ // 'NULL'  } @_ ));
 
                     my $ret;
 
@@ -341,7 +355,7 @@ package SQL::Bibliosoph; {
                     }
 
                     my ($val, $count);
-                    my $md5 = md5_hex( join ('', $name, map { $_ // 'NULL'  } @_ ));
+                    my $md5 = md5_hex( join (',', $name, map { $_ // 'NULL'  } @_ ));
                     my $md5c = $md5 . '_count';
 
                     if (! $cfg->{force} ) {
